timer_checher_printf

+++ Time_cheker/lunch_timer.sh
@@ -8,7 +8,25 @@ else
 	then
 	echo "TIMES TOO LARGE"
 	else
-		gcc -Wall -Wextra -Werror -D SIZE=$1 -D TIMES=$2 libft.a time_checker.c && ./a.out
-		rm -f a.out
+		if [[ $3 = "split" ]]
+		then
+			gcc -Wall -Wextra -Werror -D SIZE=$1 -D TIMES=$2 -D FUNCTION=1 time_checker.c libft.a && ./a.out
+			rm -f a.out
+		elif [[ $3 = "printf" ]]
+		then
+			gcc -Wall -Wextra -Werror -D SIZE=$1 -D TIMES=$2 -D FUNCTION=2 time_checker.c libft.a && ./a.out
+			rm -f a.out
+		elif [[ $3 = "gnl" ]]
+		then
+			gcc -Wall -Wextra -Werror -D SIZE=$1 -D TIMES=$2 -D FUNCTION=3 time_checker.c libft.a && ./a.out
+			rm -f a.out
+		elif [[ $3 = "all" ]]
+		then
+			gcc -Wall -Wextra -Werror -D SIZE=$1 -D TIMES=$2 -D FUNCTION=0 time_checker.c libft.a && ./a.out
+			rm -f a.out
+		else
+			gcc -Wall -Wextra -Werror -D SIZE=$1 -D TIMES=$2 -D FUNCTION=4 time_checker.c libft.a && ./a.out
+			rm -f a.out
+		fi
 	fi
 fi
